change kb_module to files module, change vs_module to kb_module for easier read plus bugs to fix the new topic and subject
