+++ users_module.py
@@ -584,12 +584,21 @@ def load_available_shared_owned_vector_stores(user_id):
 
     return accessible_vectorstores
 
+def remove_duplicates_from_vector_stores(vectorstores):
+    seen_vs_ids = set()
+    unique_vectorstores = []
+    for vs in vectorstores:
+        if vs['vs_id'] not in seen_vs_ids:
+            seen_vs_ids.add(vs['vs_id'])
+            unique_vectorstores.append(vs)
+    return unique_vectorstores
 
 def vectorstore_selection_interface(user_id):
     """
     Display Streamlit interface for vectorstore selection.
     """
-    available_vectorstores = load_available_shared_owned_vector_stores(user_id)
+    vectorstores = load_available_shared_owned_vector_stores(user_id)
+    available_vectorstores = remove_duplicates_from_vector_stores(vectorstores)
 
     st.subheader("Select knowledge base for the Bot:")
     st.write(f"Current loaded Knowledge Base: **:blue[{st.session_state.current_model}]**")
